Remove intent for save/load of dynamic components
By my current thinking, this'll just end up having to be lazy loaded
by having a vector of playlists and libaries passed to each playlist
after loading.
In that case, it'd be easier just to re-add the sources.

Therefore, I'll hold out on this until usage shows how/whether to
implement it.